Moved weight trait and minors
